+++ src/gulp/spec.js
@@ -1,30 +1,20 @@
 /**
- * A Gulp task for running specs using Karma.
+ * A Gulp task for running specs with Jest
  *
- * To use this in a gulpfile.js:
+ * To use this in gulpfile.js
  *
  *    import SpecTask from 'carbon-factory/lib/gulp/spec';
  *    gulp.task('test', SpecTask());
  *
+ *
  * You can also pass a hash of options to customise the task:
  *
  *    import SpecTask from 'carbon-factory/lib/gulp/spec';
- *
+ 
  *    var opts = {
- *      path: "/src/*.js",
- *      specs: "/src/*.spec.js",
- *      preProcessors: [ 'eslint', 'browserify' ],
- *      specPreProcessors: [ 'browserify' ],
- *      ignoreCoverage: [ '/path/to/ignore' ],
- *      reporters: ['progress'],
- *      coverageReporters: [{ type: 'text-summary' }, { type: 'html' }],
- *      coverage: {
- *        statements: 100,
- *        branches: 100,
- *        functions: 100,
- *        lines: 100
- *      },
- *      configFile: 'karma.conf.js'
+ *      jestConfig: {
+ *        watch: false
+ *      }
  *    }
  *
  *    gulp.task('test', SpecTask(opts));
@@ -34,249 +24,70 @@
  *
  *    gulp test
  *
- * By default this will automatically run in PhantomJS and will watch for any
+ * By default this will automatically run in JestCLI and will watch for any
  * changes. You can supply additional arguments to run in alternative modes:
  *
- * == To run in watch mode with coverage reports:
- *
- *    gulp test --coverage
- *
- * == To perform a single run with coverage:
+ * == To perform a single run which fails on coverage and linting (useful for CI)
  *
  *    gulp test --build
- *
- * == To run the tests in a specific browser:
- *
- *    gulp test -b chrome
- *
- * == To run the test across all browsers:
- *
- *    gulp test -b all
  */
 
-import './utils/fs-patch';
-import gutil from 'gulp-util';
-import S from 'string';
-import yargs from 'yargs';
-import karma from 'karma';
-import istanbul from 'browserify-istanbul';
-import babelify from 'babelify';
-import tsify from 'tsify';
-import fs from 'fs';
+var gulp = require('gulp');
+var jest = require('jest-cli');
+var yargs = require('yargs');
+var lint = require('./lint').default;
+var gutil = require('gulp-util');
+var fs = require('fs');
+
+
+// Config Options https://facebook.github.io/jest/docs/configuration.html
+var baseJestConfig = {
+  preset: __dirname + "./../../jest.conf.json"
+}
 
 var argv = yargs.argv;
-var Server = karma.Server;
 
-export default function(opts) {
-  var options = opts;
+export default function(options) {
+  var opts = options;
 
   return function(done) {
-    var opts = options || {};
-    // the js files to test (everything except __spec__ files)
-    var path = opts.path || '/src/***/**/!(__spec__).js';
-    // the specs
-    var specs = opts.specs || '/src/***/**/__spec__.js';
-    // an array of paths to ignore from coverage reports
-    var ignoreCoverage = opts.ignoreCoverage || [ '**/node_modules/**', '**/__spec__.js' ];
-    // an array to specify what kind of reporters should karma generate
-    var reporters = opts.reporters || ['progress'];
-    // an array to specify what kind of reporters type should karma generate
-    var coverageReporters = opts.coverageReporters || [{ type: 'text-summary' }, { type: 'html' }];
-    // which preprocessors the js files should run through
-    var preProcessors = opts.preProcessors || [ 'browserify' ];
-    // which preprocessors the spec files should run through
-    var specpreProcessors = opts.specpreProcessors || [ 'browserify' ];
-    // where to find the karma config file
-    var configFile = opts.configFile || __dirname + '/karma.conf.js';
-    // defaults the coverage thresholds
-    var coverageThreshold = Object.assign({}, {
-      statements: 100,
-      branches: 100,
-      functions: 100,
-      lines: 100
-    }, opts.coverage);
-    // eslintThreshold for failing build
-    var eslintThreshold = opts.eslintThreshold || null;
-    // Stop build if above eslintThreshold
-    var stopAboveEslintThreshold = !!eslintThreshold;
-    // coverage thresholds for each file
-    var coverageThresholdEachFile = opts.coverageEachFile || {};
-    // if using typescript
-    var typescript = opts.typescript || false;
-    // where the gulp task was ran from
-    var originPath = process.cwd();
-    // array of modules to apply babel transforms to
-    var babelTransforms = opts.babelTransforms || [];
+    // use custom config supplied via opts, or use our base config
+    var config = opts.jestConfig || baseJestConfig,
+        cliOptions = { watch: true, onlyChanged: true, config: config }
 
-    // check if eslintrc file exists, if not then prompt dev to create one
-    fs.stat(originPath + '/.eslintrc', function(err, stat) {
+    fs.stat(process.cwd() + '/.babelrc', function(err, stat) {
       if (err == null) { return }
       if (err.code == 'ENOENT') {
-        gutil.log(gutil.colors.red("Cannot find '.eslintrc' file"));
-        gutil.log(gutil.colors.white("Create an '.eslintrc' file in the root of your project and add the following code:\n{\n  \"extends\": \"./node_modules/carbon-factory/.eslintrc\"\n}"));
+        gutil.log(gutil.colors.red("Cannot find '.babelrc' file"));
+        gutil.log(gutil.colors.white("Create a '.babelrc' file in the root of your project and add the following code:\n{\n  \"extends\": \"./node_modules/carbon-factory/.babelrc\"\n}"));
         process.exit();
       }
     });
 
-    if (!argv['skip-eslint']) {
-      preProcessors.unshift('eslint');
-    }
-
-    // prefix the paths with where the gulp task was ran from so the files can
-    // be found from the correct location
-    var src = originPath + path,
-        specSrc = originPath + specs,
-        babelOptions = {
-          babelrc: false, // do not use babelrc files in gulp task
-          extends: originPath + '/node_modules/carbon-factory/.babelrc', // manually set babelrc for gulp task
-          env: {
-            test: {
-              auxiliaryCommentBefore: "istanbul ignore next",
-              plugins: [ [ "istanbul", { exclude: ignoreCoverage } ] ]
-            }
-          }
-        };
-
-    if (babelTransforms.length) {
-      var only = "^((?!node_modules).";
-
-      babelTransforms.forEach((module) => {
-        only += "|(node_modules\/" + module + ")";
-      });
-
-      only += ")*$";
-
-      babelOptions.global = true;
-      babelOptions.only = new RegExp(only);
-    } else {
-      // only babelify files in the src directory
-      babelOptions.ignore = /node_modules/;
-    }
-
-    // default configuration for the spec build
-    var config = {
-      // all the files that should be included
-      files: [ '__spec_helper__/*.js', src, { pattern: specSrc, watched: false, included: true, served: true } ],
-      // the karma config file
-      configFile: configFile,
-      // the preprocessors to run the files through
-      preprocessors: {},
-      // which browsers to run the specs through
-      browsers: [ 'PhantomJS' ],
-      // options for browserify
-      browserify: {
-        debug: true,
-        // lookup paths for any imported files
-        paths: [ originPath + '/src' ],
-        // configure any transforms for browserify
-        transform: [
-          babelify.configure(babelOptions)
-        ],
-        configure: function(bundle) {
-          bundle.on('prebundle', function() {
-            bundle.external('react/addons');
-            bundle.external('react/lib/ReactContext');
-            bundle.external('react/lib/ExecutionEnvironment');
-          });
-        },
-        cache: {},
-        packageCache: {}
-      },
-      // what kind of reporters should karma generate
-      reporters: reporters,
-      // auto watch for any changes to rerun specs
-      autoWatch: true,
-      // only run the specs once
-      singleRun: true,
-      // setup config for coverage
-      coverageReporter: {
-        dir: originPath + '/coverage',
-        reporters: coverageReporters,
-        check: {
-          global: coverageThreshold,
-          each: coverageThresholdEachFile
-        }
-      },
-      // config for eslint
-      eslint: {
-        stopOnError: false,
-        stopOnWarning: false
-      },
-      // adds additional opts for chrome browser - remembers prefs for console
-      customLaunchers: {
-        Chrome_dev: {
-          base: 'Chrome',
-          flags: ['--user-data-dir=./.browser-preferences']
-        }
-      }
-    };
-
-    // Report tests slower than value
-    if (argv['report-slow']) {
-      config.reportSlowerThan = 100;
-    }
-
-    if (typescript) {
-      config.browserify.plugin = [ tsify ];
-    }
-
-    // tie the preprocessors to the relevant sources
-    config.preprocessors[src] = preProcessors;
-    config.preprocessors[specSrc] = specpreProcessors;
-
-    if (argv.b == 'all') {
-      // if `gulp -b all` then run through all browsers
-      config.browsers = [ 'PhantomJS', 'Chrome_dev', 'Firefox', 'Safari' ];
-    } else if (argv.b) {
-      // if `gulp -b [browser]` then use the browser supplied
-      config.browsers = [ S(argv.b).capitalize().s ];
-
-      // use the custom chrome launcher
-      if (config.browsers[0] === 'Chrome') {
-        config.browsers = [ config.browsers[0] + '_dev' ];
-      }
-    }
-
-    // if coverage is enabled
-    if (argv.build || argv.coverage) {
-      process.env.NODE_ENV = 'test';
-
-      config.reporters.push('coverage');
-      config.browserify.transform.push(
-        istanbul({
-          ignore: ignoreCoverage
-        })
-      );
-    }
-
     if (argv.build) {
-      // if `gulp --build` then use single run mode
-      config.autoWatch = false;
-      // stop on lint failures in build mode
-      config.eslint.stopOnError = !stopAboveEslintThreshold;
-      // error threshold above which build fails
-      config.eslint.errorThreshold = eslintThreshold;
-      // Fail build above error threshold
-      config.eslint.stopAboveErrorThreshold = stopAboveEslintThreshold;
-      // disable source maps in build mode
-      config.browserify.debug = false;
-    } else {
-      // default to watch mode
-      config.singleRun = false;
+      cliOptions = { config: config }
     }
 
-    var server = new Server(config, function(status) {
-      if (status) {
-        gutil.log(gutil.colors.red('UI Tests FAILED'));
-      } else {
-        gutil.log(gutil.colors.green('UI Tests SUCCEEDED'));
-      }
+    if (argv.runInBand) {
+      cliOptions.runInBand = true;
+    }
 
+    // TODO: Can we pass more arguments here to jestCli
+    // https://github.com/facebook/jest/blob/master/packages/jest-cli/src/cli/index.js
+    jest.runCLI(cliOptions, '.', function(results) {
       done();
 
-      process.exit(status);
-    });
+      if (results.success) {
+        gutil.log(gutil.colors.green('UI Tests SUCCEEDED'));
 
-    server.start();
-  };
+        if (argv.build) {
+          gulp.task('lint', lint(opts));
+          gulp.start('lint');
+        }
+      } else {
+        gutil.log(gutil.colors.red('UI Tests FAILED'));
+        process.exit(1);
+      }
+    });
+  }
 }